stats indicator in all game scenes and ship camera glith workaround
