+++ TGC.Group/Model/Scenes/GameAbstractScene.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Microsoft.DirectX;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -6,6 +7,8 @@ using System.Threading.Tasks;
 using TGC.Core.Input;
 using TGC.Group.Model.Elements;
 using TGC.Group.Model.Items;
+using TGC.Group.Model.UI;
+using TGC.Group.Model.Utils;
 using TGC.Group.TGCUtils;
 
 namespace TGC.Group.Model.Scenes
@@ -20,15 +23,18 @@ namespace TGC.Group.Model.Scenes
     }
     abstract class GameAbstractScene : Scene
     {
-        public static GameState InitialGameState = new GameState(new Player.Character());
+        public static GameState InitialGameState => new GameState(new Player.Character());
         CustomSprite cursor, aim, hand;
         string dialogName, dialogDescription;
         public delegate void TransitionCallback(GameState gameState);
         public GameState GameState { get; set; }
 
+        protected StatsIndicators statsIndicators;
+
         public GameAbstractScene(GameState gameState) : base()
         {
             this.GameState = gameState;
+            InitStatsIndicator();
         }
         public override void Render()
         {
@@ -44,6 +50,16 @@ namespace TGC.Group.Model.Scenes
         {
             this.GameState = gameState;
             return this;
+        }
+        private void InitStatsIndicator()
+        {
+            Vector2 niceOffset = new Vector2(Screen.Width * 0.1f, Screen.Height * 0.05f);
+            int baseX0 = (int)(niceOffset.X);
+            int baseY0 = (int)(Screen.Height - (StatsIndicators.OxygenMeterSize + niceOffset.Y));
+
+            statsIndicators = new StatsIndicators(baseX0, baseY0);
+
+            this.statsIndicators.init();
         }
     }
 }
